Fixed a regression with 1.463
In 1.463, publisher UI was changed to do drop-down menu selection like
builders, and as a result, the code that leads to the rendering of
config.jelly has changed a bit.

More specifically, previously all the publisher config fragments are
synchronously rendered, but it switched to on-demand lazy rendering
when the user actually clicks to activate a specific publisher.

Because of lazy rendering, some of the variables that were implicitly
available is no longer available, and ${d} was one of them.

I'll fix this in the core to keep those plugins working, but
in the mean time I also fixed this plugin to use ${descriptor},
which is more blessed.