Re-order initialization of intermediate directory
In a recent commit, the initialization of the intermediate
directory got moved to after launcher decoration checks if it
should do anything. Usually this doesn't matter, because it does
apply changes, but in the case of C#, cov-build does not need to
be run, so the method exits early. This means that the
intermediate directory object is never created, causing a null
pointer exception.