Split up analyze and commit steps
The new C# defect import is required to be run only once, after
analysis, but before commit. The previous analyze/commit process
consisted of looping through the streams configured for a job, and
running analysis if necessary, then committing. The new process is
to first run all of the analyses, then run the C# defect import,
and then run all of the commits.