Make sure that platform is detected correctly for decorated launchers
When the ant action is run, it called Launcher.isUnix() to determine
which executable to use (.bat or not), and whether to add "cmd /C"
around the ant command line call.

Previously, CoverityLauncherDecorator did not delegate this method,
so Launcher.isUnix() would be called instead of the appropriate
delegated method. This delegation is now implemented correctly.