Fix:77998,77968
Added getWsVersion() to CIMInstance, which allows us to know the ws version of a particular cim instance. This is key if we are going to support both versions at the same time.

On v9, the drop down menu for "Languages" (under configurations for a particular job) will only display "ALL" and we will import all checkers.
On v6, the behavior remains the same.