Enabling Cov-Manage-History when there is no snapshots should not fail.
With the changes in this commit, the cov-manage-history command will make a webservice call to check if there is a snapshot for given stream and if there is not, then the cov-mange-history will not be run and continue with other configured post-build actions. Thus, the build will not be marked as build failure due to lack of snapshot.

Fix: 67320