Added caching of the user permissions check result
This will bypass performing a new check when an instance has already been checked. The result will be set when performing a check from either the job configuration (Check Configuration button) or the check at beginning of a build. The result will be reset if the global configuration is updated.

Fix: 105657